test(kernel): convert trajectory feed BDD to Gherkin + pytest-bdd
Replace the informal docstring-BDD test_kernel_trajectory_feed_bdd.py with
a real tests/features/kernel_trajectory_feed.feature plus pytest-bdd step
definitions, covering all 5 scenarios (create_temp pending->ready feed,
non-platform markdown rejection, pending-never-fed, incremental atom dedup,
and demo kernel offline distill). Adds pytest-bdd to dev deps.
