+++ pyproject.toml
@@ -52,7 +52,7 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-dev = ["pytest>=7", "pytest-timeout>=2", "pytest-asyncio>=0.21", "requests>=2"]
+dev = ["pytest>=7", "pytest-timeout>=2", "pytest-asyncio>=0.21", "pytest-bdd>=7", "requests>=2"]
 
 [project.urls]
 Homepage = "https://github.com/SkillNerds/xskill"
